requirements.txt
aniso8601==6.0.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.3
Flask-JWT==0.3.2
Flask-JWT-Extended==3.18.2
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
mysqlclient==1.4.2.post1
pluggy==0.11.0
py==1.8.0
PyJWT==1.4.2
PyMySQL==0.9.3
python-decouple==3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
